# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: Fantastic Coffee (decaffeinated)
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
security:
  - bearerAuth: []
components:   
  securitySchemes:
    bearerAuth:     # arbitrary name for the security scheme
      description: "arbitrary name for the security scheme"       
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    ID: 
      title: General ID format 
      description: "Specifies the format of a unique identifier (UUID)."
      type: string 
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
      example: 3e7b82c1-8f66-4f54-9e57-7baa8a0ea8e3
               
               # 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
      readOnly: true
...
    Like:
      title: Like of a post
      description: |-
        like from a user to a photo/post of another user 
      type: object
      properties:
        # identifier of the post to which the like belongs  
        post_id:
          $ref: "#/components/schemas/ID"
        
        like_id:
          $ref: "#/components/schemas/ID"
        # id of the user who does the comment/like, this will help us identify from whom the comment/like is
        user_id: 
          $ref: "#/components/schemas/ID"
...
paths:
...
  /session: 
    post:
      tags: ["login"]
      summary: Logs in the user 
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin 
      requestBody:
        description: User details
        content:
          application/json: 
            schema:
              type: object 
              properties:
                name:
                  type: string 
                  example: Maria 
                  pattern: 'ˆ.*?$'
                  minLength: 3 
                  maxLength: 16
        required: true 
      responses:
        '201':
          description: User log-in action successful 
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    $ref: "#/components/schemas/ID"
                    # change here if you decide to use an integer # or any other type of identifier
                    #type: string
                    #example: 1a2b3c4d5e6f7A8B9C0D1E2F3A4B5C6D
                    
  
  # LIKE PHOTO
  # here we want to add a like to a post 
  /likes/:
    post:
      security:
        - bearerAuth: []
      tags:
        - Posts
      operationId: likePhoto
      summary: Add a like to a post
      description: |- 
        an user is trying to add a like to a post of another user 
      
          
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Like"}
      responses: 
        "201":
          # in this case by created we mean that we add a like to the array of likes of a post 
          description: Like successfully created 
          # we need a response body given that we are creating a new resources
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Like"} 
              # means that the responses of the request is in a json format and indicates how it's organizes, what it has ecc...
              
        '400':
          description: Bad request. Invalid user data.
            
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error.
  
  # UNLIKE PHOTO
  # here we want to remove a like that we previously added to a post 
  /likes/{likeid}:
    delete: 
      security:
        - bearerAuth: []
      tags:
        - Posts
      operationId: unlikePhoto
      summary: Remove a like from a post 
      description: |- 
        an user want to remove e like that previously add to a post of another user 
      parameters:
        - name: likeid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true 
        
      responses:
        # use status code 204 instead of 200, because with 200 there is response body with additional information which can be misleading in the context of a DELETE operation. with 204 indicates that the deletion operation was successful
        
        "204":
          description: |-
            specify if the like of an user post was correctly deletated 
        '400':
          description: Bad request. Invalid user data.
            
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: Resource not found 
          
        "500":
          description: Internal server error 
  
  # COMMENT PHOTO        
  #here we want to add a comment to a post (it could be to the self post or to post of another user)
...
  /users/{userid}/posts/{postid}:
    delete:
      security: 
        - bearerAuth: []
      tags:
        - Profile 
        - Users
      operationId: deletePhoto
      summary: Remove post from profile
      description: |-
        user want to remove post from his own profile 
      
      parameters:
        - name: userid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
          
        - name: postid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
      
      responses:
        "204":
          description: |- 
            specify if the post in the user profile was correctly deletated
            
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'

        "404":
          description: Resource not found
          
        "500":
          description: Internal server error 
          
  # FOLLOW USER        
  # user want to  follow an user 
...