# service/api/api-photo.go
package api

import (
	"database/sql"
	"encoding/json"
	"errors"
	"github.com/julienschmidt/httprouter"
	"net/http"
)

func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
	w.Header().Set("Content-Type", "application/json")

	// get from the path the userid and handle error
	userid := ps.ByName("userid")
	if userid == "" {
		rt.baseLogger.Warning("the userid in the path is empty")
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	// Authentication
	authorized := Authentication(w, r, userid)
	if authorized == false {
		return
	}

	// create a variable post to parse the data passed in the request body
	var p Post

	// Read and parse the JSON data from the request body into a Post object.
	err := json.NewDecoder(r.Body).Decode(&p)
	if err != nil {
		rt.baseLogger.WithError(err).Warning("Invalid JSON format")
		w.WriteHeader(http.StatusBadRequest)
		_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "bad request body"})
		return
	}

	// defer closing the request body
	defer r.Body.Close()

	// Generate a new unique ID for the like
	postid := generateUniqueID()
	newpost := Post{
		User_ID:  userid,
		Post_ID:  postid,
		Uploaded: p.Uploaded,
		Image:    p.Image,
	}

	// Save the new post in the database
	err = rt.db.UploadPhoto(newpost.PostToDataBase())
	if err != nil {
		rt.baseLogger.WithError(err).Warning("Error saving post into database")
		w.WriteHeader(http.StatusInternalServerError)
		_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "Internal server error"})
		return
	}

	// Encode the newpost to a variable before writing it to the response writer
	response, err := json.Marshal(newpost)
	if err != nil {
		rt.baseLogger.WithError(err).Warning("Error encoding post for response")
		w.WriteHeader(http.StatusInternalServerError)
		_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "Internal server error"})
		return
	}

	// Write the encoded response to the response writer
	w.WriteHeader(http.StatusCreated)
	_, err = w.Write(response)
	if err != nil {
		rt.baseLogger.WithError(err).Warning("Error writing response")
		w.WriteHeader(http.StatusInternalServerError)
		_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "Internal server error"})
		return
	}
}

func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
	w.Header().Set("Content-Type", "application/json")

	postid := ps.ByName("postid")
	if postid == "" {
		// Handle the case when "likeid" is not present in the request.
		rt.baseLogger.Warning("the postid in the path is empty")
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	userid, _ := rt.db.GetUserIDForPost(postid)
	// Authentication
	authorized := Authentication(w, r, userid)
	if authorized == false {
		return
	}

	// Remove from the database the post given the postid
	err := rt.db.DeletePhoto(postid)
	if err != nil {
		// check it the err is of the same type of sql.ErrNoRows
		if errors.Is(err, sql.ErrNoRows) {
			// Resource not found
			rt.baseLogger.WithError(err).Warning("Post not found in the database")
			w.WriteHeader(http.StatusNotFound)
			_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "Post not found in the database"})
			return
		}

		// Handle other errors
		rt.baseLogger.WithError(err).Warning("Error removing post from the database")
		w.WriteHeader(http.StatusInternalServerError)
		_ = json.NewEncoder(w).Encode(JSONErrorMsg{Message: "Internal server error"})
		return
	}

	w.WriteHeader(http.StatusNoContent)
}